recipe-api/requirements.txt
django>=5.1.4
python-dotenv
openai
moviepy
ffmpeg-python
transformers
torch
numpy
whisper
Pillow


